Kruskal-Wallis Test
to decide whether the population distributions are identical without assuming them to follow the normal distribution.

The ANOVA (and t-test) is explicitly a test of equality of means of values. The Kruskal-Wallis (and Mann-Whitney) can be seen technically as a comparison of the mean ranks, testing whether the probability that a random observation from each group is equally likely to be above or below a random observation from another group.