typical for bioinformatics
https://ibb.co/CPkY0tc
As it's an unsupervised problem, the PCA projections can be used for an unsupervised clustering technique
If it was supervised - 
Take the PCA projections and build a classifier off of that. Use a scree/elbow plot to determine the number of components to use in your classifier. Below is some R code I wrote to teach this.
https://www.reddit.com/r/bioinformatics/comments/chpgoz/hey_guys_this_pca_plot_clearly_indicates_two/